Trousers: Add trousers to the repo
This commit adds trousers to the repo and makes two changes. The first is that
tinylogin is no longer available and replaced with busybox. The second is that
patches no longer reside in the same directory as the recipe file. They either
need to go in files or a folder named ${PN}-${PV}